+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Security
 
-## [0.16.2] - 2019-11-18
+## [0.16.2] - 2019-11-22
 
 ### Added